AD-136 Moved signInWithNostr() method
